requirements.txt
psutil
py-cpuinfo
customtkinter
requests